Q&Achatbot/requirements.txt
langchain-groq
langchain
python-dotenv
langchain_community
streamlit
openai
langchain_openai